% Auto-generated: do not edit by hand
\name{''DashMasonry}

\alias{''DashMasonry}

\title{DashMasonry component}

\description{

}

\usage{
''DashMasonry(children=NULL, className=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Contents of grid.}

\item{className}{Character. Style class of the component.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

